azure-savings-report/requirements.txt
# Azure Authentication
azure-identity>=1.15.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Azure Data Explorer (optional - for ADX data source)
azure-kusto-data>=4.3.0

# Azure Storage (optional - for storage exports)
azure-storage-blob>=12.19.0

# Visualization & Reporting
plotly>=5.18.0
jinja2>=3.1.2
openpyxl>=3.1.2  # Excel export

# HTTP requests (for Retail Prices API)
requests>=2.31.0
httpx>=0.25.0

# Configuration
pyyaml>=6.0.1

# CLI
typer>=0.9.0
rich>=13.7.0
